Refactor into multiple classes
